Moved sample_ruptures in hazardlib
